Generate GVN during HIR creation.
The title seems innocuous, however having a unique identifier for each
value in the HIR from the beginning should allow refactoring subsequent
passes to use this identifier to refer to elements.

Hopefully, this means shifting away from the current graph traversal
and incremental transformation approach, to using tables in the Context
and updating them in-place.

Then, once no progress is made in updating the Context any longer, a
finalization pass should apply all transformations to the graph at
once.