Introduce hir::Tree, a self-contained representation.
The idea of the Tree is to have a self-contained representation of the
whole expression, while preserving random-access to individual nodes.

The Tree is basically an ECS applied to the compiler, and the idea is
to amortize the allocations by recycling old Trees when they are no
longer useful.

The Tree is immediately leveraged through the use of the Context for
implementing Nested Entity Fetching and Type Unification. No more
walking and rebuilding the whole Abstract Tree from scratch at each
iteration!

Instead the Context contains a work-queue of items to process as well
as links between the items. Whenever an item is "resolved", linked
items are added to the queue automatically, so they can be
re-evaluted, and potentially resolved too.

The system is still a bit wonky:
- Initial seeding of the Context is shaky.
- Links may not be setup properly yet.
- Type Unification is very rudimentary.

Furthermore, there is now a dual representation of types in the HIR:
- hir::Type, used in the Tree, is only meaningful within its tree.
- hir::TypeDefinition, stand-alone, and requiring allocations.

And of course, the very principle use for hir::Tree could be used for
ssa::ControlFlowGraph, with identical gains. It may also help weed out
some uses of hir::TypeDefinition, which would be quite helpful...

Lots of further work, thus, but a single integration test is failing
and this patch is already very big.