Introduce com::CoreId
It is unlikely that IDs over 32-bits will be necessary within the compiler as
having 4 billion of any given entity would simply require a massive amount of
memory, even if it were generated code through use of generics.

As a result, it makes sense to introduce a space-optimized and tested CoreId
type to help build special-purpose IDs rather than having a swarm of ad-hoc
implementations scattered left and right.