Introduce ast::Module and ast::Tree, self-contained representations
The idea is similar to the motivation for hir::Tree, this yields a
self-contained representation of the whole module or expression which
is both cache-friendly (a few vectors) and easily accomodates new
elements.

The Module is an ECS used to record all the items defined within a
given module (file), regardless of their level of nesting.

The Tree only represents a single Expression or Function, and indeed
a Module will typically contain multiple Trees.

Modules and Trees can, in the future, easily be recycled.

There are still a handful of temporary memory allocations when parsing,
scattered uses of Vec for building slices. Those should, later, be
migrated to some SmallVec structure so that small slices (say, of at
most 8 or 16 items) would not require any allocation at all.