Move Store and MultiStore traits to basic module.
Since they pair up with the Id struct, it is only natural that they be
defined in the same module than it is, so as to be usable as widely as
possible.

Also, take the opportunity to make the Id to use configurable, though
with a suitable default, so that custom Id (hir::ExpressionId, for
example) can still be used with the trait.