Prepare JaggedArray and JaggedHashMap for concurrency.
Using a Snapshot on a different thread is made easier by removing the
lifetime parameter which tied it to the main instance.

Also, the many uses of unsafe have been reviewed, with a good faith
effort at ensuring their correctness, and a number of interfaces have
been foolproofed. Still not quite ready for prime-time, though.

JaggedArray is still not Sync, so cannot be written to concurrently
from multiple threads at once, however it should be close to be ready
for reading from multiple threads even while another is writing.