Introduce HIR Repository and Module, revamp SIR Graph
Continuing on the ECS trend, introduce HIR Repository and Module.

A Repository will gather all items program-wide, while a Module will gather all
items of a particular module (ah!).

Both are implemented as ECS, and HIR IDs related to items using a simplistic
scheme: 1 billion for Tree, 1 billion for Module, 1 billion for Repository and
the leftover billion for special needs.

This effectively deprecates TypeDefinition, and removes most long-lived
discrete allocations, paving the way toward allocation-less cruise speed.

Notably, this allows for the removal of mem::heap::{DynArray, Ptr} and allows
actually "freezing" the definitions, ensuring they can safely be shared across
multiple threads.

As a ripple effect, the SIR and Interpreter, which use HIR types, had to be
migrated to the new structure, which meant also migrating the ControlFlowGraph
to the ECS trend.

A lots of clones died as a result \o/