Remove prototypes
The idea behind prototypes was that the definition of an item could be
referring to an item not yet defined, and therefore a two-pass approach
was necessary: first gather the prototypes, register them, and then
only complete the definitions.

Unfortunately, prototypes do not solve this problem fully. Or more
specifically, while they seem to solve the problem now, they do not
solve it in a future where the prototype itself would need to refer to
other names; such as when generics will enter play.

The introduction of strongly typed IDs, however, means that a solution
to the problem is now as easy as first registering the names and then
perform a second pass to parse the definitions.

Names were already introduced ahead of time with Module, so it is just
a matter of removing prototypes and refactoring the code base.