Handle adding/removing sequence elements in MultiTable
The original implementation of MultiTable aimed at getting something
off the ground as quickly as possible, and therefore was limited to
only ever appending new sequences, without any possibility to add or
remove elements from existing sequences afterward.

The new implementation is a tad more complicated, however it handles
adding and removing elements to existing sequences and does so
efficiently regardless of the number of other sequences in the table.

The key insight is that multiple vectors can be used for storage. The
layout consists of:

- An index: mapping an integer (0..len) to a SequenceIndex which
  indexes the sequence in two dimensions: table and offset.
- An array of arrays: where each array contains sequences up to twice
  as long as the previous array, leading to capacities of 1, 2, 4, 8,
  16, ... elements.

Moving a sequence from one sub-array to another, or removing a sequence
altogether, may require filling the gap it leaves behind, which is
handled by switching in the last sequence of the sub-array and updating
its mapping in the index.

Arrays are cache-friendly, and their memory allocations is amortized,
leading to a non-allocating cruise-speed with efficient O(1) handling
of sequences.