Minimal support for implementations
Change
------

Introduces parsing and semantic support for implementing an interface.

The parsing support is complete, and the semantic support is sufficient
to resolve the interface and extended type as well as properly scoping
the functions defined inside.

Next
----

Handle cross-referencing interfaces and types, so as to be able to:

- Handle sub-typing relationships.
- Cast from type to interface.
- Perform interface method calls from type.

Next Next
---------

Even then it still will be incomplete:

- The SIR does not represent virtual calls.
- The interpreter does not handle virtual calls.