Minimal support for method calls
Change
------

Implement proper resolution for method calls, based on the type of the
receiver.

A method call may resolve to:
- A function, in which case the receiver is the first argument.
- An interface method, providing the interface is in scope.
- An extension function, based on the type of the receiver.

It is important to note that this resolution is purely SCOPE based, and
that no pruning is performed based on whether the number of arguments
matches, or the call may possibly type checks.

The name must be unambiguous within the scope, or a qualified function
call must be used to disambiguate it -- such as "Iface::func(record)".