Reintroduce Implicit casts
Change
------

During earlier refactorings, Implicit casts had been temporarily
removed to allow the refactoring to proceed.

This changes reintroduces them so that when the Type Unification
computes an Action of type Cast(target) the cast is actually applied
instead of causing a panic -- in a number of circumstances.

Limitation
----------

For now, implicit casts only allow going from a Record to an Enum.

Left out
--------

While the Tree should be correctly updated, no effort has been made to
update the links between GVNs. It is unclear whether such updates are
actually necessary -- though Murphy would hint that they are.