Introduce Implicit::ToInt to cast to interface
Change
------

Add the cast and implement the Action.

Context
-------

Implicit casts are represtend in the HIR explicitly, although introduced by
the compiler implicitly and not appearing in the source code.

The intent is to have 3 implicit casts:

- ToEnum: representing a cast from a Record to an Enum.
- ToInt: representing a cast from any Type to an Interface.
- ToStruct: representing a cast from a Tuple to a Record.

The latter being much more limited in scope -- only available when the type of
the record is explicit in the context.

Hum
---

At the moment, though, those casts are not actually introduced by the compiler.