Introduce methods in SIR, and in the interpreter
Change
------

Introduce method calls in the SIR, lowering of method calls from HIR to
SIR in the SSA pass, and handle method calls in interpreter.

A slew of small bug fixes, both in syn and sem is fixed along the way.

Tests
-----

So, I didn't create tests for the small fixes:

- There is no high-level :imp parsing test, apparently.
- There is no high-level type unification finalization test, either.

There are some more pressing issues, though.

Next
----

Make Extensions and Implementations work for built-in types.

Next Next
---------

And brace for tuples, which requires canonicalization of tuples.

And should allow getting rid of "local" Type IDs in sir::Graph, which
were a nice little hack, but simply do not work when the local Type ID
of one graph is propagated (through a tuple) to a call into the next
graph which has different mappings.

Instead, tuples should be canonicalized, and TupleId should appear in
Module and Registry to uniquely identify a tuple, which implies
rewriting Tuples in Modules/Trees when canonicalizing...

Perhaps starting with canonicalizing Types would be a good start,
there's many duplicates which is rather unnecessary.