Strip away range information from hir::Tuple
Change
------

The names of the hir::Tuple fields stored both the Identifier and
the range at which it appeared in the source code.

The latter is unnecessary for processing, and makes canonicalization
impossible: the ranges can be stripped for comparison, but then they
are no longer accurate.

Rather than hold onto inaccurate/misleading information, strip it away.

They are kept only for function arguments, as the order of arguments
matter.

Next
----

Inching closer to tuple canonicalization...