Canonicalize Types
Change
------

Canonicalize Types, such that within a Repository or Module, a given
TypeId always refers to the same type.

This adds extra processing, at the cost of saving quite a bit of memory
whenever types are mentioned more than 2x beyond their definition,
which is expected to be the case.

Although the actual motivation is to make Tuples canonical, which is
necessary to be able to write Extensions and Implementations for
tuples.

Note that types are not canonicalized within a Tree, this is because
type inference takes place within a Tree, and any type with a type
variable still unresolved cannot be canonicalized.

Next
----

Fill in the holes? There's some bits and pieces here and there which
may require some fiddling.

For example, the fact that named tuples do not yet sort the names...

Next Next
---------

Actually implementing Extensions and Implementations on Tuple types!