Canonicalize Types in Tree
Change
------

Rewrite the types in a Tree to refer to the canonical types of the
Repository -- and fix stuff left and right as the Repository was not
used until now.

This eliminates the need for copying types into the sir Graph.

Next
----

Still not sorting those tuples before insertion...

Next Next
---------

Implementing extensions and interfaces on tuples is close!